requirements.txt
absl-py==1.4.0
accelerate==0.30.0
annotated-types==0.7.0
beautifulsoup4==4.13.3
bitsandbytes==0.44.1
brotlipy==0.7.0
bypy==1.8
cachetools==5.3.1
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
clip==1.0
cryptography @ file:///croot/cryptography_1677533068310/work
# Editable Git install with no remote (dassl==0.6.3)
-e /home/zy/Dassl.pytorch
deepspeed==0.15.3
dill==0.3.9
docker-pycreds==0.4.0
einops==0.5.0
entrypoints==0.3
filelock==3.12.2
flake8==3.7.9
fsspec==2024.3.1
ftfy==6.1.1
future==0.18.3
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
hjson==3.1.0
huggingface-hub==0.25.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.8.0
info-nce-pytorch==0.1.4
isort==4.3.21
Jinja2==3.1.4
joblib==1.3.1
littleutils==0.2.2
Markdown==3.4.4
MarkupSafe==2.1.3
mccabe==0.6.1
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.1.0
multiprocess==0.70.17
networkx==3.1
ninja==1.11.1.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ogb==1.3.6
opencv-python==4.6.0.66
outdated==0.2.2
packaging==24.0
pandas==2.0.2
peft==0.7.1
Pillow==9.4.0
platformdirs==4.3.6
protobuf==4.23.4
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.5.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==2.1.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests @ file:///croot/requests_1682607517574/work
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.0
scipy==1.10.1
sentry-sdk==2.17.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sympy==1.12
tabulate==0.9.0
tb-nightly==2.14.0a20230803
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.15.2
torch==1.13.0
torchaudio==0.12.1
torchvision==0.14.0
tqdm==4.65.0
transformers==4.37.1
triton==2.3.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1686163155763/work
wandb==0.18.5
wcwidth==0.2.6
Werkzeug==2.3.6
wilds==1.2.2
yacs==0.1.8
yapf==0.29.0
zipp==3.16.2
